KATs passing: includes temporary (?) workarounds

+++ crypto_kem/hqc-rmrs-128/clean/code.c
@@ -21,10 +21,13 @@
  * @param[out] em Pointer to an array that is the tensor code word
  * @param[in] m Pointer to an array that is the message
  */
-void PQCLEAN_HQCRMRS128_CLEAN_code_encode(uint8_t *em, const uint8_t *m) {
+void PQCLEAN_HQCRMRS128_CLEAN_code_encode(uint64_t *em, const uint8_t *m) {
     uint8_t tmp[VEC_N1_SIZE_BYTES] = {0};
+    // should m be stored as bytes or words?
 
+    // operates on bytes
     PQCLEAN_HQCRMRS128_CLEAN_reed_solomon_encode(tmp, m);
+    // can we make this encode words?
     PQCLEAN_HQCRMRS128_CLEAN_reed_muller_encode(em, tmp);
 
 }
@@ -37,9 +40,10 @@ void PQCLEAN_HQCRMRS128_CLEAN_code_encode(uint8_t *em, const uint8_t *m) {
  * @param[out] m Pointer to an array that is the message
  * @param[in] em Pointer to an array that is the code word
  */
-void PQCLEAN_HQCRMRS128_CLEAN_code_decode(uint8_t *m, const uint8_t *em) {
+void PQCLEAN_HQCRMRS128_CLEAN_code_decode(uint8_t *m, const uint64_t *em) {
     uint8_t tmp[VEC_N1_SIZE_BYTES] = {0};
 
+    // can we make this decode words?
     PQCLEAN_HQCRMRS128_CLEAN_reed_muller_decode(tmp, em);
     PQCLEAN_HQCRMRS128_CLEAN_reed_solomon_decode(m, tmp);
