+++ crypto_kem/hqc-rmrs-128/clean/code.h
@@ -10,9 +10,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void PQCLEAN_HQCRMRS128_CLEAN_code_encode(uint8_t *em, const uint8_t *message);
+void PQCLEAN_HQCRMRS128_CLEAN_code_encode(uint64_t *em, const uint8_t *message);
 
-void PQCLEAN_HQCRMRS128_CLEAN_code_decode(uint8_t *m, const uint8_t *em);
+void PQCLEAN_HQCRMRS128_CLEAN_code_decode(uint8_t *m, const uint64_t *em);
 
 
 #endif
